requirements.txt
alembic==1.16.1
sqlalchemy==2.0.41
python-dotenv==1.1.0
fastapi~=0.115.12
fastapi-users[sqlalchemy]==14.0.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
